[SCM-656] Building maven-scm-1.6 requires a native install of git. skip test if git is not in PATH
git-svn-id: https://svn.apache.org/repos/asf/maven/scm/trunk@1240126 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-scm-providers/maven-scm-providers-git/maven-scm-provider-gitexe/src/test/java/org/apache/maven/scm/provider/git/gitexe/command/checkin/GitCheckInCommandNoBranchTest.java
@@ -27,7 +27,6 @@ import org.apache.maven.scm.command.checkin.CheckInScmResult;
 import org.apache.maven.scm.command.checkout.CheckOutScmResult;
 import org.apache.maven.scm.repository.ScmRepository;
 import org.codehaus.plexus.util.FileUtils;
-import org.codehaus.plexus.util.Os;
 
 import java.io.File;
 
@@ -51,6 +50,11 @@ public class GitCheckInCommandNoBranchTest
     public void testCheckinNoBranch()
         throws Exception
     {
+        if ( !ScmTestCase.isSystemCmd( "git" ) )
+        {
+            System.out.println( "skip test which git native executable in path" );
+            return;
+        }
         File repo_orig = new File( "src/test/resources/repository_no_branch" );
         File repo = getTestFile( "target/git_copy" );
         FileUtils.deleteDirectory( repo );
